core: jwt: move helper methods to sync under submodule
